requirements.txt
-e .
beautifulsoup4
coverage
coveralls
django-extensions
flake8
flake8-bugbear
flake8-comprehensions
gitchangelog
ipython
pdbpp
pre_commit
psycopg2cffi
pytest
pytest-cov
simplejson
Sphinx
sphinx-autobuild
sphinx_rtd_theme
sqlalchemy_utils
tox
tox-factor
tox-pyenv
